Repository restructuring pt.3: remove obsolete csv_loader.py and implement transactions_csv_loader.py with enhanced data validation and upload processes. Moves functions from db_utils.py and distributes between test_transactions_manager.py and transactions_csv_loader.py.
