--- budgy/database_modules/io/csv_loader.py
@@ -1,15 +1,0 @@
-#!python3
-"""
-"""
-# Standard library imports
-
-
-# Third party imports
-
-# Custom imports
-
-# Local imports
-
-# initialize module logger
-import logging
-logger = logging.getLogger(__name__)
